gennf, gennch ; split expression with two calls to random number generator to ensure deterministic evaluation
both routines include an expression which calls two functions which
use random number generator.  Since C does not guarantee the order of
evaluation within an expression it is possible for different compilers
to invoke the functions in different orders, resulting in different
values for the expression when starting from the same seed.

This change breaks up the expressions to ensure a deterministic order.